replace CustomVideoProcessor for custom VideoSink
